+++ tests/testthat/test-qgis-configure.R
@@ -170,6 +170,97 @@ test_that("qgis_configure() works OK if cache conditions unmet", {
 
 
 
+test_that("qgis_configure() works OK if cache condition 'use_json_output' unmet", {
+  skip_if_not(has_qgis())
+  skip_if_not(
+    package_version(qgis_version(full = FALSE)) >= "3.23.0",
+    "This QGIS version does not support JSON input (needed for this test)."
+  )
+  version <- as.character(utils::packageVersion("qgisprocess"))
+  cache_data_file <- file.path(
+    rappdirs::user_cache_dir("R-qgisprocess"),
+    glue("cache-{version}.rds")
+  )
+  withr::defer(
+    saveRDS(
+      list(
+        path = qgis_path(),
+        version = qgis_version(),
+        algorithms = qgis_algorithms(),
+        plugins = qgis_plugins(),
+        use_json_output = qgis_using_json_output()
+      ),
+      cache_data_file
+    )
+  )
+
+  # user setting 'use_json_output' FALSE will trigger reconfiguration if cached value was TRUE
+  withr::local_options(list(
+    qgisprocess.use_json_input = NULL,
+    qgisprocess.use_json_output = FALSE
+  ))
+  saveRDS(
+    list(
+      path = qgis_path(),
+      version = qgis_version(),
+      algorithms = qgis_algorithms(),
+      plugins = qgis_plugins(),
+      use_json_output = TRUE
+    ),
+    cache_data_file
+  )
+  expect_message(
+    capture.output(qgis_configure(use_cached_data = TRUE)),
+    "contradict the 'use_json_output' cache value"
+  )
+  expect_false(qgis_using_json_output())
+  expect_false(qgis_using_json_input())
+
+  # absence of user settings will honor the previously set cached value FALSE
+  withr::local_options(list(
+    qgisprocess.use_json_output = NULL
+  ))
+  saveRDS(
+    list(
+      path = qgis_path(),
+      version = qgis_version(),
+      algorithms = qgis_algorithms(),
+      plugins = qgis_plugins(),
+      use_json_output = FALSE
+    ),
+    cache_data_file
+  )
+  expect_no_message(
+    capture.output(qgis_configure(use_cached_data = TRUE)),
+    message = "contradict the 'use_json_output' cache value"
+  )
+  expect_false(qgis_using_json_output())
+  expect_false(qgis_using_json_input())
+
+  # user setting 'use_json_input' TRUE will trigger reconfiguration if cached value was FALSE
+  withr::local_options(list(
+    qgisprocess.use_json_input = TRUE # expectations below need JSON input to be supported!
+  ))
+  saveRDS(
+    list(
+      path = qgis_path(),
+      version = qgis_version(),
+      algorithms = qgis_algorithms(),
+      plugins = qgis_plugins(),
+      use_json_output = FALSE
+    ),
+    cache_data_file
+  )
+  expect_message(
+    capture.output(qgis_configure(use_cached_data = TRUE)),
+    "contradict the 'use_json_output' cache value"
+  )
+  expect_true(qgis_using_json_input())
+  expect_true(qgis_using_json_output())
+})
+
+
+
 
 test_that("abort_query_version() works", {
   lines <- c("aa", "bb")